+++ src/components/UserSettings/Shortcuts/index.tsx
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 import { parseKeyCombo } from "@blueprintjs/core";
 
 import { TextField, SettingsManager } from "Components/UserSettings";
-import { camelToTitleCase } from "../../../utils";
+import { camelToTitleCase } from "../common";
 
 
 const { Provider: ShortcutsProvider, useSettings: useShortcutsSettings } = new SettingsManager<"shortcuts">();
