+++ src/components/UserSettings/common.test.ts
@@ -1,0 +1,18 @@
+import { camelToTitleCase } from "./common";
+
+
+describe("camelToTitleCase", () => {
+	const cases = [
+		["someText", "Some Text"],
+		["toggleDashboard", "Toggle Dashboard"],
+		["copyAsReference", "Copy As Reference"]
+	];
+
+	test.each(cases)(
+		"%# - %s",
+		(input, expectation) => {
+			expect(camelToTitleCase(input))
+				.toBe(expectation);
+		}
+	);
+});
