+++ src/components/UserSettings/common.tsx
@@ -15,6 +15,17 @@ const popoverProps: Partial<IPopoverProps> = {
 	popoverClassName: CustomClasses.POPOVER
 };
 
+
+/** Converts a string from camelCase to Title Case
+ * @see https://stackoverflow.com/questions/7225407/convert-camelcasetext-to-title-case-text */
+function camelToTitleCase(text: string): string {
+	const result = text.replace(/([A-Z])/g, " $1");
+	const finalResult = result.charAt(0).toUpperCase() + result.slice(1);
+
+	return finalResult;
+}
+
+
 const Row: FC = ({ children }) => <div zr-role="settings-row">{children}</div>;
 
 
@@ -363,6 +374,7 @@ const Toggle = ({ description = null, isChecked, label, onChange = undefined, ti
 
 
 export {
+	camelToTitleCase,
 	Definition,
 	NumericSelect,
 	RowCol,
